drawing: extract generic HyperExplorer base classes; mpl subclasses to backends/mpl/
Create backend-agnostic base classes HyperExplorer (he.py),
HyperSignal1D_Explorer (hse.py), and HyperImage_Explorer (hie.py) with
abstract methods for backend-specific operations. Move matplotlib
implementations to backends/mpl/mpl_he.py, mpl_hse.py, mpl_hie.py using
multiple inheritance. Old drawing/mpl_he.py, mpl_hse.py, mpl_hie.py become
one-line re-export shims. Update signal.py to import from new backend paths.
